modify get, get all and delete calls of master data disposal

+++ gdpr/src/main/java/com/kairos/controller/master_data/asset_management/DataDisposalController.java
@@ -49,7 +49,7 @@ public class DataDisposalController {
 
     @ApiOperation("get DataDisposal by id")
     @GetMapping("/data_disposal/{dataDisposalId}")
-    public ResponseEntity<Object> getDataDisposal(@PathVariable Long countryId, @PathVariable BigInteger dataDisposalId) {
+    public ResponseEntity<Object> getDataDisposal(@PathVariable Long countryId, @PathVariable Integer dataDisposalId) {
         return ResponseHandler.generateResponse(HttpStatus.OK, true, dataDisposalService.getDataDisposalById(countryId, dataDisposalId));
 
     }
